chore: refactor code
- I blame @leeerob from making me do this refactor https://x.com/leeerob/status/1810707436551991412
