+++ components/comments.tsx
@@ -2,8 +2,8 @@ import { Suspense } from 'react';
 import { auth } from '@/config/auth';
 import { db } from '@/lib/db';
 import { formatDate } from '@/lib/utils';
-import { SignInButton } from './auth/signin-button';
-import { SignOutButton } from './auth/signout-button';
+import { SignInButton } from './signin-button';
+import { SignOutButton } from './signout-button';
 import { CommentArea } from './comment-area';
 import { Avatar, AvatarFallback, AvatarImage } from './ui/avatar';
 import { Skeleton } from './ui/skeleton';
