--- components/layout/header.tsx
@@ -1,57 +1,0 @@
-import { SiGithub, SiLinkedin, SiX } from '@icons-pack/react-simple-icons';
-import { NavLink } from './nav-link';
-
-export function Header() {
-  return (
-    <header className='sticky top-0 z-50 w-full border-b shadow-md backdrop-blur'>
-      <nav className='mx-auto flex max-w-screen-lg items-center justify-between p-4'>
-        <ul className='flex gap-4'>
-          <li>
-            <NavLink href='/'>home</NavLink>
-          </li>
-          <li>
-            <NavLink href='/work'>work</NavLink>
-          </li>
-          <li>
-            <NavLink href='/blogs'>blogs</NavLink>
-          </li>
-        </ul>
-        <ul className='flex gap-4'>
-          <li className='hover:opacity-70'>
-            <a
-              href='https://x.com/_nikhilsnayak_'
-              className='dark:text-fluorescent'
-              target='_blank'
-              rel='noopener noreferrer'
-              aria-label='x.com'
-            >
-              <SiX className='size-4' />
-            </a>
-          </li>
-          <li className='hover:opacity-70'>
-            <a
-              href='https://github.com/nikhilsnayak'
-              className='dark:text-fluorescent'
-              target='_blank'
-              rel='noopener noreferrer'
-              aria-label='github'
-            >
-              <SiGithub className='size-4' />
-            </a>
-          </li>
-          <li className='hover:opacity-70'>
-            <a
-              href='https://linkedin.com/in/nikhilsnayak'
-              className='dark:text-fluorescent'
-              target='_blank'
-              rel='noopener noreferrer'
-              aria-label='linkedin'
-            >
-              <SiLinkedin className='size-4' />
-            </a>
-          </li>
-        </ul>
-      </nav>
-    </header>
-  );
-}
